delete toy task and move observation to new structure
